Radio Buttons and Action Tool listeners
I got the Line action button to be selected as the default tool. I also got it where if we click on 1 of the radio buttons, it will select the proper action button and set the current tool to the corresponding tool. I also got it to work vice versa. But... I seem to have encountered a problem since I get an error when I click on the eraser tool with the action button or radio button. It doesnt do it with the pencil tool though. So im not sure what I did to break the program when I only messed with the GUI.